Feature/export multiple files (#1)
* feat: Export multiple files

---------

+++ MotionTracking/FilesList/FilesListViewController.swift
@@ -15,6 +15,18 @@ class FilesListViewController: BaseViewController<FilesListViewModel> {
     
     // MARK: Override methode
     
+    override func setupUI() {
+        super.setupUI()
+        
+        tableView.allowsMultipleSelectionDuringEditing = true
+        navigationItem.rightBarButtonItem = editButtonItem
+        
+        let selectAllButton = UIBarButtonItem(title: "Select all", style: .plain, target: self, action: #selector(selectAllTapped))
+        let exportButton = UIBarButtonItem(barButtonSystemItem: .action, target: self, action: #selector(exportTapped))
+        let spacer = UIBarButtonItem(barButtonSystemItem: .flexibleSpace, target: self, action: nil)
+        toolbarItems = [selectAllButton, spacer, exportButton]
+    }
+
     override func setupObservers() {
         super.setupObservers()
 
@@ -35,7 +47,7 @@ class FilesListViewController: BaseViewController<FilesListViewModel> {
         // Insert row in tableview
         viewModel.inserFile
             .receive(on: DispatchQueue.main)
-            .sink(receiveValue: { [weak self]   _ in
+            .sink(receiveValue: { [weak self] _ in
                 self?.inserRows()
             })
             .store(in: &cancellable)
@@ -48,6 +60,55 @@ class FilesListViewController: BaseViewController<FilesListViewModel> {
             }
             .store(in: &cancellable)
         
+        // Show the toolbar
+        viewModel.$showToolBar
+            .removeDuplicates()
+            .receive(on: DispatchQueue.main)
+            .sink { [weak self] value in
+                self?.navigationController?.setToolbarHidden(!value, animated: true)
+            }
+            .store(in: &cancellable)
+        
+        // Enable the edit button
+        viewModel.$isEnabledEditButton
+            .removeDuplicates()
+            .sink { [weak self] value in
+                self?.toolbarItems?.last?.isEnabled = value
+            }
+            .store(in: &cancellable)
+        
+        // Show table view editing
+        viewModel.$isEditing
+            .removeDuplicates()
+            .receive(on: DispatchQueue.main)
+            .sink { [weak self] value in
+                if self?.tableView.isEditing != value {
+                    self?.tableView.setEditing(value, animated: true)
+                }
+            }
+            .store(in: &cancellable)
+        
+        // Select all row
+        viewModel.$isAllSelected
+            .removeDuplicates()
+            .receive(on: DispatchQueue.main)
+            .sink { [weak self] value in
+                self?.selectAllRow(value)
+            }
+            .store(in: &cancellable)
+        
+        viewModel.$stateSelectedButton
+            .removeDuplicates()
+            .receive(on: DispatchQueue.main)
+            .sink { [weak self] value in
+                if value {
+                    self?.toolbarItems?.first?.title = "Deselect all"
+                }else{
+                    self?.toolbarItems?.first?.title = "Select all"
+                }
+            }
+            .store(in: &cancellable)
+        
     }
     
     /**
@@ -56,10 +117,16 @@ class FilesListViewController: BaseViewController<FilesListViewModel> {
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
         if let navigationController = segue.destination as? UINavigationController,
            let viewController = navigationController.topViewController as? ExportFileViewController {
-            guard let indexPath = tableView.indexPathForSelectedRow else { return }
-            viewController.viewModel.fileTrackingEntity = viewModel.files[indexPath.row]
+            viewModel.$filesSelected.assign(to: &viewController.viewModel.$fileTrackingEntity)
         }
+
     }
+    
+    override func setEditing(_ editing: Bool, animated: Bool) {
+        super.setEditing(editing, animated: animated)
+        viewModel.isEditing = !viewModel.isEditing
+    }
+    
     
     // MARK: Private methode
     
@@ -73,6 +140,33 @@ class FilesListViewController: BaseViewController<FilesListViewModel> {
     
     private func reloadData() {
         tableView.reloadData()
+    }
+    
+    @objc func exportTapped() {
+        performSegue(withIdentifier: "ExportFileSegue", sender: nil)
+    }
+    
+    @objc func selectAllTapped(sender: UIBarButtonItem) {
+        viewModel.isAllSelected = !viewModel.isAllSelected
+    }
+    
+    private func selectAllRow(_ value: Bool) {
+        for section in 0..<tableView.numberOfSections {
+            for row in 0..<tableView.numberOfRows(inSection: section) {
+                let indexPath = IndexPath(row: row, section: section)
+                if !value {
+                    tableView.deselectRow(at: indexPath, animated: false)
+                } else {
+                    tableView.selectRow(at: IndexPath(row: row, section: section), animated: false, scrollPosition: .none)
+                }
+            }
+        }
+        filesSelected()
+    }
+    
+    private func filesSelected() {
+        viewModel.filesSelected = tableView.indexPathsForSelectedRows?
+            .compactMap { viewModel.files[$0.row] } ?? []
     }
 
 }
@@ -92,13 +186,23 @@ extension FilesListViewController: UITableViewDataSource, UITableViewDelegate {
     }
     
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        tableView.deselectRow(at: indexPath, animated: true)
+        filesSelected()
+        if !tableView.isEditing  {
+            performSegue(withIdentifier: "ExportFileSegue", sender: nil)
+            tableView.deselectRow(at: indexPath, animated: true)
+        }
+    }
+    
+    func tableView(_ tableView: UITableView, didDeselectRowAt indexPath: IndexPath) {
+        filesSelected()
     }
     
     func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
-        viewModel.removeFile(index: indexPath.row)
+        if editingStyle == .delete {
+            viewModel.removeFile(index: indexPath.row)
+        }
     }
-    
+
 }
